# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Shop API",
        "description": "Documentation officielle de l'API Shop",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Serveur local"
        }
    ],
    "paths": {
...
        "/api/sales": {
            "get": {
                "tags": [
                    "Sales"
                ],
                "description": "Récupérer toutes les ventes de l'utilisateur connecté",
                "operationId": "419ebb078c83262b05f7c639ae8b2c21",
                "responses": {
                    "200": {
                        "description": "Liste des ventes récupérée"
                    },
                    "204": {
                        "description": "Aucune commande enregistrée"
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/sales/annullee": {
            "get": {
                "tags": [
                    "Sales"
                ],
                "description": "Récupérer les ventes annulées",
                "operationId": "f4b41e3fd84e3b305453f3ffbe68fbc5",
                "responses": {
                    "200": {
                        "description": "Liste des ventes annulées récupérée"
                    },
                    "201": {
                        "description": "Aucune commande annulée"
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/sales/expediee": {
            "get": {
                "tags": [
                    "Sales"
                ],
                "description": "Récupérer les ventes expédiées",
                "operationId": "6ceaf27fedfeba87ec3157fdfa53b00a",
                "responses": {
                    "200": {
                        "description": "Liste des ventes expédiées récupérée"
                    },
                    "204": {
                        "description": "Aucune commande expédiée"
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/sales/en_cours": {
            "get": {
                "tags": [
                    "Sales"
                ],
                "description": "Récupérer les ventes en cours",
                "operationId": "c7d423b99f0558f012684047a3b6a595",
                "responses": {
                    "200": {
                        "description": "Liste des ventes en cours récupérée"
                    },
                    "204": {
                        "description": "Aucune commande en cours"
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/sales/store": {
            "post": {
                "tags": [
                    "Sales"
                ],
                "description": "Créer une nouvelle vente et décrémenter le stock des produits",
                "operationId": "d86adba5204a6d4d8d3a43afdceae00a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "products"
                                ],
                                "properties": {
                                    "products": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "product_id",
                                                "quantity"
                                            ],
                                            "properties": {
                                                "product_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "quantity": {
                                                    "type": "integer",
                                                    "example": 2
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Commande créée avec succès"
                    },
                    "404": {
                        "description": "Produit introuvable"
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}